servicedirectory/src/requirements.txt
Django==1.5.1
djangorestframework==2.3.5
jsonschema==2.0.0
pymongo==2.5

